Arcade Machine Chance Event
Event idea is when the player runs into the arcade machine the player goes into the arcade and turns into a sprite that start a mini game for the player to play on the Arcade. Currently only has the game functions. Will work on

A cutscene for when the player activates the chance events.

Have controls work with the arcade machine model to play the game like a real machine

Edit Code to have the screen turned off unless interacted with by the player